pyproject.toml
[project]
name = "Underground_Arkanoid"
version = "1.0.0"
description = "Modificated Arkanoid"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pygame-ce>=2.5.3",
]
